+++ .gitignore
@@ -1,3 +1,3 @@
 .DS_Store
 .vscode
-*.py
+builds/*
